requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
psycopg2-binary
python-dotenv==1.0.1
mail-parser==3.15.0
dnspython==2.4.2
msal==1.24.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
requests==2.31.0
boto3==1.28.44